requirements.txt
beautifulsoup4==4.6.0
EasyProcess==0.2.3
Pillow==5.0.0
pygame==1.9.3
pyscreenshot==0.4.2
pytesseract==0.1.7
SimpleCV==1.3
asyncio==3.4.3
nltk==3.2.5
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
idna==2.6
requests==2.18.4
urllib3==1.22
